+++ build.gradle
@@ -1,0 +1,44 @@
+plugins {
+    id 'java'
+    id 'org.jetbrains.intellij' version '0.4.16'
+    id 'org.jetbrains.kotlin.jvm' version '1.3.60'
+}
+
+group pluginGroup
+version pluginVersion
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+}
+
+sourceSets {
+    main.kotlin.srcDirs += 'src/main/kotlin'
+}
+
+patchPluginXml {
+    version = project.version
+}
+
+intellij {
+    version ideaVersion
+    type ideaType
+    pluginName pluginName
+    updateSinceUntilBuild = false
+    downloadSources Boolean.valueOf(sources)
+
+    plugins = ['JavaScriptLanguage', 'JavaScriptDebugger', 'NodeJS']
+}
+
+compileKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
+
+compileTestKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
