+++ gradle.properties
@@ -1,0 +1,13 @@
+# Available idea versions:
+# https://www.jetbrains.com/intellij-repository/releases
+# https://www.jetbrains.com/intellij-repository/snapshots
+
+ideaVersion = 2019.2
+ideaType = IU
+sources = true
+
+pluginGroup = io.plugin.tsnode
+pluginName = Run Configuration for TypeScript
+pluginVersion = 2020.03.04.0
+
+kotlin.code.style=official
