+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Mar 06 11:42:30 CET 2020
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.2.1-all.zip
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
