+++ src/main/kotlin/io/plugin/tsnode/action/DebugTsAction.kt
@@ -1,6 +1,6 @@
 package io.plugin.tsnode.action
 
-import icons.TsIcons
+import io.plugin.tsnode.icons.TsIcons
 
 class DebugTsAction : TsAction(TsIcons.TypeScriptDebug)
 {
