+++ src/main/kotlin/io/plugin/tsnode/execution/TsConfigurationFactory.kt
