+++ src/main/kotlin/io/plugin/tsnode/execution/TsConfigurationType.kt
