+++ src/main/kotlin/io/plugin/tsnode/execution/TsProgramRunner.kt
