+++ src/main/kotlin/io/plugin/tsnode/lib/TsLog.kt
