+++ src/main/kotlin/io/plugin/tsnode/runner/_ConfigurationEditor.kt
@@ -1,4 +1,4 @@
-package io.plugin.base.runner
+package io.plugin.tsnode.runner
 
 import com.intellij.openapi.options.SettingsEditor
 import com.intellij.openapi.project.Project
