+++ src/main/resources/icons/typescript-debug@16.png
